Make sense of render thread #1
What is now called GLResource should live only on the render thread. Instead of
the main thread creating GLResources and queueing them for build/destruct on the
render thread, the Renderer should generate RenderResources and return handles
to these.

This initial commit introduces the RenderResource and RenderResourceManager
which are both intended to replace / be merged with the GLResource and
GLResourceManager once all GLResources have been transformed to RenderResources.

Currently only textures and frame buffers are made into RenderResources.